requirements.txt
httpx
beautifulsoup4
python-dotenv